+++ src/Ext/BuildingType/Body.h
@@ -22,6 +22,7 @@ public:
 		ValueableVector<BuildingTypeClass*> PowersUp_Buildings;
 		ValueableIdxVector<SuperWeaponTypeClass> SuperWeapons;
 
+		Valueable<double> PowerPlant_DamageFactor;
 		ValueableVector<BuildingTypeClass*> PowerPlantEnhancer_Buildings;
 		Valueable<int> PowerPlantEnhancer_Amount;
 		Nullable<float> PowerPlantEnhancer_Factor;
@@ -84,7 +85,6 @@ public:
 		Valueable<int> Overpower_KeepOnline;
 		Valueable<int> Overpower_ChargeWeapon;
 
-		Valueable<bool> PackupSound_PlayGlobal;
 		Valueable<bool> DisableDamageSound;
 		Nullable<float> BuildingOccupyDamageMult;
 		Nullable<float> BuildingOccupyROFMult;
@@ -93,11 +93,10 @@ public:
 		NullableIdx<VocClass> BunkerWallsUpSound;
 		NullableIdx<VocClass> BunkerWallsDownSound;
 
-		Valueable<double> PowerPlant_DamageFactor;
-
 		ExtData(BuildingTypeClass* OwnerObject) : Extension<BuildingTypeClass>(OwnerObject)
 			, PowersUp_Owner { AffectedHouse::Owner }
 			, PowersUp_Buildings {}
+			, PowerPlant_DamageFactor { 1.0 }
 			, PowerPlantEnhancer_Buildings {}
 			, PowerPlantEnhancer_Amount { 0 }
 			, PowerPlantEnhancer_Factor { 1.0 }
@@ -146,7 +145,6 @@ public:
 			, BarracksExitCell {}
 			, Overpower_KeepOnline { 2 }
 			, Overpower_ChargeWeapon { 1 }
-			, PackupSound_PlayGlobal { false }
 			, DisableDamageSound { false }
 			, BuildingOccupyDamageMult {}
 			, BuildingOccupyROFMult {}
@@ -154,7 +152,6 @@ public:
 			, BuildingBunkerROFMult {}
 			, BunkerWallsUpSound {}
 			, BunkerWallsDownSound {}
-			, PowerPlant_DamageFactor { 1.0 }
 		{ }
 
 		// Ares 0.A functions
@@ -191,7 +188,7 @@ public:
 	static bool LoadGlobals(PhobosStreamReader& Stm);
 	static bool SaveGlobals(PhobosStreamWriter& Stm);
 
-	static void PlayBunkerSound(BuildingClass const* pThis, bool bUp = false);
+	static void PlayBunkerSound(BuildingClass const* pThis, bool buildUp = false);
 
 	static int GetEnhancedPower(BuildingClass* pBuilding, HouseClass* pHouse);
 	static bool CanUpgrade(BuildingClass* pBuilding, BuildingTypeClass* pUpgradeType, HouseClass* pUpgradeOwner);